+++ javax-kt/src/main/java/me/panpf/javaxkt/util/Formatx.kt
@@ -18,6 +18,7 @@
 
 package me.panpf.javaxkt.util
 
+import me.panpf.javax.util.FormatTotalTimeConfig
 import me.panpf.javax.util.Formatx
 import me.panpf.javaxkt.lang.orZero
 
@@ -159,6 +160,11 @@ inline fun Int.formatShortFileSize(): String = Formatx.shortFileSize(this.toLong
 
 /* ******************************************* totalTime *******************************************/
 
+
+/**
+ * Returns the total time of formatting that can be displayed
+ */
+inline fun Long.formatTotalTime(config: FormatTotalTimeConfig): String = Formatx.totalTime(this, config)
 
 /**
  * Returns the total time of formatting that can be displayed