+++ javax/src/test/java/me/panpf/javax/test/util/FormatxTest.java
@@ -177,8 +177,15 @@ public class FormatxTest {
         assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 2), "1d 1h 23m");
         assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 3), "1d 1h");
         assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 4), "1d");
-        assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 5), "0s");
-        assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 6), "0s");
+        assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 5), "0d");
+        assertEquals(Formatx.totalTime(1000 * 60 * 60 * 24 + 1000 * 60 * 60 + (1000 * 60 * 23) + (1000 * 23) + 467, 6), "0d");
+
+        assertEquals(Formatx.totalTime(0, 0), "0s");
+        assertEquals(Formatx.totalTime(0, 1), "0s");
+        assertEquals(Formatx.totalTime(0, 2), "0m");
+        assertEquals(Formatx.totalTime(0, 3), "0h");
+        assertEquals(Formatx.totalTime(0, 4), "0d");
+        assertEquals(Formatx.totalTime(0, 5), "0d");
     }
 
     @Test